requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
Pillow==9.4.0
requests==2.28.2
urllib3==1.26.14
